requirements.txt
google-cloud-aiplatform==1.59.0
google-cloud-bigquery==3.25.0
streamlit==1.36.0
google-cloud-aiplatform[reasoningengine,langchain]
langchain-google-vertexai==1.0.4
langfuse==2.39.1